refactor(T1432): Float→Decimal migration + reports-extended split + quick fixes
Schema (P1 — banking precision):
- TimeEntry.hours: Float → Decimal @db.Decimal(5,2)
- TimeEntryTemplateItem.hours: Float → Decimal @db.Decimal(5,2)
- Task.estimatedHours, Task.actualHours: Float → Decimal @db.Decimal(5,2)
- 16 API routes + 5 services updated with Number() conversion

Component split:
- reports-extended.tsx: 1089 → ~52 lines (7 sub-report components extracted)
- DateRangeProps extracted to shared report-types.ts

Quick fixes:
- generate-reset-token: remove redundant requireMinRole call
- backup.sh: lock file moved from /tmp to BACKUP_ROOT
- lib/api-response.ts: add Decimal→number serialization helper

Closes #1432
