+++ app/components/reports/report-table-audit.tsx
@@ -12,10 +12,8 @@ import { exportCSV } from "./report-charts";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
-export interface DateRangeProps {
-  from: string;
-  to: string;
-}
+import type { DateRangeProps } from "./report-types";
+export type { DateRangeProps } from "./report-types";
 
 interface AuditEntry {
   id: string;
