+++ lib/api-response.ts
@@ -1,3 +1,4 @@
+import { Prisma } from "@prisma/client";
 import { NextResponse } from "next/server";
 
 export interface ApiResponse<T = unknown> {
@@ -8,10 +9,27 @@ export interface ApiResponse<T = unknown> {
 }
 
 /**
+ * JSON replacer that converts Prisma Decimal values to plain JS numbers so
+ * that API responses return numeric values (not Decimal strings) for
+ * Decimal fields (e.g. hours, estimatedHours, actualHours).
+ */
+function decimalReplacer(_key: string, value: unknown): unknown {
+  if (Prisma.Decimal.isDecimal(value)) {
+    return (value as Prisma.Decimal).toNumber();
+  }
+  return value;
+}
+
+/**
  * Returns a successful JSON response with consistent { ok, data } shape.
+ * Prisma Decimal fields are automatically serialized as numbers.
  */
 export function success<T>(data: T, status = 200): NextResponse<ApiResponse<T>> {
-  return NextResponse.json({ ok: true, data }, { status });
+  const body = JSON.stringify({ ok: true, data }, decimalReplacer);
+  return new NextResponse(body, {
+    status,
+    headers: { "Content-Type": "application/json" },
+  });
 }
 
 /**
